Fix: consumer retryable 적용, exist 먼저 처리 Refactor: document 가격 관련 fieldtype 변경, kafka서버값 application으로 공통 사용

+++ elastic-search/src/main/java/store/_0982/elasticsearch/application/dto/GroupPurchaseDocumentInfo.java
@@ -12,7 +12,7 @@ public record GroupPurchaseDocumentInfo(
         Integer maxQuantity,
         String title,
         String description,
-        Integer discountedPrice,
+        Long discountedPrice,
         String status,
         Instant startAt,
         Instant endAt,
