+++ elastic-search/src/main/java/store/_0982/elasticsearch/presentation/dto/GroupPurchaseDocumentRequest.java
@@ -12,7 +12,7 @@ public record GroupPurchaseDocumentRequest(
         Integer maxQuantity,
         String title,
         String description,
-        Integer discountedPrice,
+        Long discountedPrice,
         String status,
         Instant startAt,
         Instant endAt,
